+++ src/Command/Install/CreateAdminCommand.php
@@ -1,0 +1,52 @@
+<?php
+/**
+ * This file is part of me-cms.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       2.26.0
+ */
+namespace MeCms\Command\Install;
+
+use Cake\Console\Arguments;
+use Cake\Console\ConsoleIo;
+use Cake\Console\ConsoleOptionParser;
+use MeCms\Command\AddUserCommand;
+use MeTools\Console\Command;
+
+/**
+ * Creates an admin user
+ */
+class CreateAdminCommand extends Command
+{
+    /**
+     * Hook method for defining this command's option parser
+     * @param ConsoleOptionParser $parser The parser to be defined
+     * @return ConsoleOptionParser
+     */
+    protected function buildOptionParser(ConsoleOptionParser $parser)
+    {
+        $parser->setDescription(__d('me_cms', 'Creates an admin user'));
+
+        return $parser;
+    }
+
+    /**
+     * Creates an admin user
+     * @param Arguments $args The command arguments
+     * @param ConsoleIo $io The console io
+     * @return null|int The exit code or null for success
+     */
+    public function execute(Arguments $args, ConsoleIo $io)
+    {
+        $command = new AddUserCommand;
+        $command->initialize();
+
+        return $command->execute(new Arguments([], ['group' => 1, 'verbose' => $args->getOption('verbose')], []), $io);
+    }
+}