+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # 2.x branch
 ## 2.26 branch
 ### 2.26.0
-* uses `addPlugin()` to load plugins. Updated for CakePHP 3.6;
+* `InstallShell` has been replaced with console commands. Every method of the
+    previous class is now a `MeCms\Command\Install` class;
 * `UserShell` has been replaced with console commands. Every method of the
-    previous class is now a `Command` class;
+    previous class is now a `MeCms\Command` class;
+* uses `addPlugin()` to load plugins. Updated for CakePHP 3.6;
 * the table `$cache` property is now protected. Added `AppTable::getCacheName()`
     method to get the cache configuration name used by the table. It can also
     returns the names of the associated tables;